pyproject.toml
[project]
name = "data-engineering-hw"
version = "0.1.0"
description = "Проект по курсу data engineering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
